Refactor config import and fix pylint issues in numba_dpex/core/utils/kernel_launcher.py
Refactor config import of numba_dpex/core/descriptor.py

Refactor config import in numba_dpex/decorators.py

Refactor config import in numba_dpex/tests/misc/test_warnings.py

Refactor config import in tests._helper.py
